[feat] myPageViewModel merge
